rename Buffer.h to BufferModule.h
